test: scenario: el rewards: move utils to helpers
