+++ test/0.8.9/oracle/accounting-oracle-deploy.test.js
@@ -6,6 +6,7 @@ const {
   updateLocatorImplementation,
   deployLocatorWithDummyAddressesImplementation,
 } = require('../../helpers/locator-deploy')
+const { calcReportDataHash, getReportDataItems } = require('../../helpers/reportData')
 
 const {
   SLOTS_PER_EPOCH,
@@ -45,35 +46,6 @@ const EXTRA_DATA_FORMAT_LIST = 1
 
 const EXTRA_DATA_TYPE_STUCK_VALIDATORS = 1
 const EXTRA_DATA_TYPE_EXITED_VALIDATORS = 2
-
-function getReportDataItems(r) {
-  return [
-    r.consensusVersion,
-    +r.refSlot,
-    r.numValidators,
-    r.clBalanceGwei,
-    r.stakingModuleIdsWithNewlyExitedValidators,
-    r.numExitedValidatorsByStakingModule,
-    r.withdrawalVaultBalance,
-    r.elRewardsVaultBalance,
-    r.lastWithdrawalRequestIdToFinalize,
-    r.finalizationShareRate,
-    r.isBunkerMode,
-    r.extraDataFormat,
-    r.extraDataHash,
-    r.extraDataItemsCount,
-  ]
-}
-
-function calcReportDataHash(reportItems) {
-  const data = web3.eth.abi.encodeParameters(
-    [
-      '(uint256,uint256,uint256,uint256,uint256[],uint256[],uint256,uint256,uint256,uint256,bool,uint256,bytes32,uint256)',
-    ],
-    [reportItems]
-  )
-  return web3.utils.keccak256(data)
-}
 
 function encodeExtraDataItem(itemIndex, itemType, moduleId, nodeOperatorIds, keysCounts) {
   const itemHeader = hex(itemIndex, 3) + hex(itemType, 2)
