+++ test/scenario/execution_layer_rewards_after_the_merge.test.js
@@ -9,6 +9,7 @@ const { waitBlocks, setBalance } = require('../helpers/blockchain')
 const { deployProtocol } = require('../helpers/protocol')
 const { setupNodeOperatorsRegistry } = require('../helpers/staking-modules')
 const { SLOTS_PER_FRAME } = require('../helpers/constants')
+const { calcReportDataHash, getReportDataItems } = require('../helpers/reportData')
 
 const NodeOperatorsRegistry = artifacts.require('NodeOperatorsRegistry')
 
@@ -33,33 +34,6 @@ const makeAccountingReport = ({ refSlot, numValidators, clBalanceGwei, elRewards
   extraDataHash: ZERO_HASH,
   extraDataItemsCount: 0,
 })
-function calcReportDataHash(reportItems) {
-  const data = web3.eth.abi.encodeParameters(
-    [
-      '(uint256,uint256,uint256,uint256,uint256[],uint256[],uint256,uint256,uint256,uint256,bool,uint256,bytes32,uint256)',
-    ],
-    [reportItems]
-  )
-  return web3.utils.keccak256(data)
-}
-function getReportDataItems(r) {
-  return [
-    r.consensusVersion,
-    +r.refSlot,
-    r.numValidators,
-    r.clBalanceGwei,
-    r.stakingModuleIdsWithNewlyExitedValidators,
-    r.numExitedValidatorsByStakingModule,
-    r.withdrawalVaultBalance,
-    r.elRewardsVaultBalance,
-    r.lastWithdrawalRequestIdToFinalize,
-    r.finalizationShareRate,
-    r.isBunkerMode,
-    r.extraDataFormat,
-    r.extraDataHash,
-    r.extraDataItemsCount,
-  ]
-}
 
 contract('Lido: merge acceptance', (addresses) => {
   const [
